Consertado erros de compilacao de scene.cc, tava faltando umas "->" no lugar de "."
git-svn-id: http://xp-dev.com/svn/horus_eye/trunk@34 863e7ce6-8861-447e-8dcb-03987b876380

+++ src/framework/animation.h
@@ -14,10 +14,10 @@ class Animation {
   public:
     Animation(float fps, ...);
     ~Animation() { }
-    
+
     void set_fps(float fps) { fps_ = fps; }
-    int fps() { return fps_; }
-    
+    float fps() { return fps_; }
+
     int get_current_frame() { return current_frame_; }
     void Update(float delta_t);
